+++ cxx/test/test_prufer.cpp
@@ -9,7 +9,7 @@
 TEST_CASE("prufer: networkx_example_parent")
 {
     const std::vector<int> pruf {3, 3, 3, 4};
-    const int n = int(pruf.size() + 2);
+    const size_t n = pruf.size() + 2;
     std::vector<int> parent (n, -1);
     prufer2parent(n, pruf.data(), parent.data());
 
@@ -21,7 +21,7 @@ TEST_CASE("prufer: networkx_example_parent")
 TEST_CASE("prufer: networkx_example_edges")
 {
     const std::vector<int> pruf {3, 3, 3, 4};
-    const int n = int(pruf.size() + 2);
+    const size_t n = pruf.size() + 2;
 
     std::vector<int>
         head (n-1, -1),
@@ -29,7 +29,7 @@ TEST_CASE("prufer: networkx_example_edges")
 
     prufer2edges(n, pruf.data(), head.data(), tail.data());
     std::set<std::set<int>> edges;
-    for (int i = 0; i < n-1; i++)
+    for (size_t i = 0; i < n-1; i++)
         edges.insert({head[i], tail[i]});
         
     const std::set<std::set<int>>
@@ -42,14 +42,14 @@ TEST_CASE("prufer: networkx_example_edges")
 TEST_CASE("prufer: paper")
 {
     const std::vector<int> c {2, 4, 0, 1, 3, 3};
-    const int n = int(c.size() + 2);
+    const size_t n = c.size() + 2;
     std::vector<int>
         head (n-1, -1),
         tail (n-0, -1);
 
     prufer2edges(n, c.data(), head.data(), tail.data());
     std::set<std::set<int>> edges;
-    for (int i = 0; i < n-1; i++)
+    for (size_t i = 0; i < n-1; i++)
         edges.insert({head[i], tail[i]});
 
     const std::set<std::set<int>>
